requirements.txt
amqp==5.3.1
annotated-types==0.7.0
billiard==4.2.1
boto3==1.40.30
botocore==1.40.30
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
datadispatch==1.0.0
durationpy==0.10
gitdb==4.0.12
GitPython==3.1.43
google-auth==2.40.3
idna==3.10
jmespath==1.0.1
kombu==5.5.4
kubernetes==33.1.0
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gitlab==6.2.0
python-json-logger==3.3.0
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.42
tasks==2.8.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
vulture==2.14
wcwidth==0.2.13
websocket-client==1.8.0
